# src/main/java/br/com/bookstoreconsumer/adapters/input/dto/AuthorRequest.java
package br.com.bookstoreconsumer.adapters.input.dto;

import br.com.bookstoreconsumer.core.domain.Author;

public record AuthorRequest(String name) {

    public static AuthorRequest toRequest(Author author) {
        return new AuthorRequest(author.getName());
    }
}